// frontend/src/app/api/fn/publico-destinos/publico-listar-destinos.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { Dificultad } from '../../models/dificultad';
import { Mes } from '../../models/mes';
import { PaginaDestinoTarjeta } from '../../models/pagina-destino-tarjeta';
import { Presupuesto } from '../../models/presupuesto';
import { Slug } from '../../models/slug';
import { TramoDuracion } from '../../models/tramo-duracion';

export interface PublicoListarDestinos$Params {

/**
 * Slugs de tipos de aventura (OR). Repetible `?tipo=a&tipo=b`.
 */
  tipo?: Array<Slug>;
  region?: Slug;
  pais?: Slug;
  dificultad_min?: Dificultad;
  dificultad_max?: Dificultad;
  mes?: Mes;

/**
 * Tramos de duración (OR).
 */
  duracion?: Array<TramoDuracion>;

/**
 * Niveles de presupuesto (OR).
 */
  presupuesto?: Array<Presupuesto>;
  orden?: 'nombre' | 'dificultad_asc' | 'dificultad_desc';

/**
 * Número de página (base 1). Fuera de rango → 404 pagina_fuera_de_rango.
 */
  pagina?: number;
}

export function publicoListarDestinos(http: HttpClient, rootUrl: string, params?: PublicoListarDestinos$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginaDestinoTarjeta>> {
  const rb = new RequestBuilder(rootUrl, publicoListarDestinos.PATH, 'get');
  if (params) {
    rb.query('tipo', params.tipo, {"style":"form","explode":true});
    rb.query('region', params.region, {});
    rb.query('pais', params.pais, {});
    rb.query('dificultad_min', params.dificultad_min, {});
    rb.query('dificultad_max', params.dificultad_max, {});
    rb.query('mes', params.mes, {});
    rb.query('duracion', params.duracion, {"style":"form","explode":true});
    rb.query('presupuesto', params.presupuesto, {"style":"form","explode":true});
    rb.query('orden', params.orden, {});
    rb.query('pagina', params.pagina, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginaDestinoTarjeta>;
    })
  );
}

publicoListarDestinos.PATH = '/api/v1/publico/destinos';